feat: 더미데이터 insert 및Chart(react)화면과 연결

+++ jpa/src/main/java/com/example/jpa/domain/Chart.java
@@ -19,14 +19,14 @@ public class Chart {
     @Column(nullable = false)
     private String basicDate; //날짜
     @Column(nullable = false) 
-    private int impCnt; //노출수
+    private Long impCnt; //노출수
     @Column(nullable = false)
-    private int clickCnt; //클릭수
+    private Long clickCnt; //클릭수
     @Column(nullable = false)
     private double convCnt; //전환율(클릭수 / 노출수)
     @Column(nullable = false)
-    private int sellCost; //판매금액
+    private Long sellCost; //판매금액
     @Column(nullable = false)
-    private int adSpend; //광고비
+    private Long adSpend; //광고비
 
 }
